+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 ## [5.52.1](https://github.com/plivo/plivo-dotnet/tree/v5.52.1) (2026-05-26)
+**Feature - Profile API DBA field support**
+- Added Doing Business As (DBA) field support to Profile API
+
 **Feature - Expose sub_account_name on RentedNumber resource**
 - Added `SubAccountName` property to the `RentedNumber` resource, surfacing the subaccount's name alongside the existing `SubAccount` (auth_id) on rented number listing and get APIs